+++ CHANGELOG.md
@@ -25,6 +25,7 @@ With this change, methods `.split_sessions()` & `.stop_only()` have been removed
 * Added class method `quick_start()`. Start a menu with it's default settings only adding the pages
 * Added property `menu_type`. Displays the menu type that was set in the constructor
 * Added method `.randomize_embed_colors()`. Selects a random color for all embeds added to the menu
+* Added method `.version_info()`. A simple shortcut to function `reactionmenu.version_info()`
 ##### ViewMenu Only
 * Added the `persist` kwarg to `ViewButton`. This prevents link buttons from being disabled/removed when the menu times out or is stopped so they can remain clickable